+++ src/tr2/game/objects/general/puzzle_hole.c
@@ -94,7 +94,7 @@ void __cdecl PuzzleHole_Collision(
     }
 
     if (g_Inv_Chosen == NO_OBJECT) {
-        Inv_Display(INV_KEYS_MODE);
+        Inv_DisplayKeys(item->object_id);
         if (g_Inv_Chosen == NO_OBJECT && g_Inv_KeyObjectsCount > 0) {
             return;
         }